+++ samples/stopwatch-coroutines/src/main/java/com/instacart/formula/stopwatch/StopwatchRenderModel.kt
@@ -1,0 +1,6 @@
+package com.instacart.formula.stopwatch
+
+data class StopwatchRenderModel(
+    val timePassed: String
+)
+